Initial commit of Tiger interpreter.
This is an interpreter for Appel's Tiger language
written in pure OCaml. Right now it compiles and
seems to give correct answers to simple expression.

Functions are not yet implemented. Just for testing
we have implemented a kind of REPL loop, but it doesn't
really work as a loop in the usual sense since you have
to write ^D each time you want an expression evaluated.
We will see if we can improve on this later.