Initial function support.
User defined functions seems to work. More
extensive testing is needed. The interpreter (Tiger_eval)
implements nested functions using a display. The maximum
static nesting depth is computed when compiling (Tiger_compile)
and then a display of that size is kept throughout computation.
For the details look in [call] in Tiger_eval.